mcshredder based memc proxy test suite
This system has two sub-suites:
- stability
- performance

This suite makes use of pre-made containers and an externalized control
script (nodectrl) to manage virtual instances running a proxy and
backend nodes. A single mcshredder instance is then run, which iterates
through all of the test scenarios.

The goal of the stability suite is to flush out race conditions, memory
leaks, slow crashes, and make subtle problems "look larger" so they are
noticed. It exercises code in situations with bandwidth constraints,
packet loss, constant configuration reloads, etc.

The goal of the performance suite is to be highly sensitive to
performance changes. The suite itself must run as efficiently as
possible and scale up enough to make very small gain/losses in performance
measurable repeatably.

Performance tests measure the timing of every request and report into
histograms and charts. They run many individual tests ramping between
low load to high load to characterize:
- latency/cpu deltas from low to high load
- performance consistency vs throughput dropoff during overload.
- pathological performance or non-linear outliers.

It also allows overriding which specific subtests to run and for how
long, ie; if adding a new "fault" type test, run it for 5 hours instead
of just 30 seconds.

Planning on allowing it to detect out of spec tests and halt. Scripts
are used to analyze the output. Might generate graphical web data if I
get around to it.