output a total histogram at end of each perf run
per-second will lose some data at the end of the run, and I have to
summarize the run in a script. will add back a mode for per-second to
test for slowdown-during-a-run conditions later.

now we do the run, then output all of the data. reduces the overhead of
the test a bit as well.